// PassportProjectInAngular/src/app/shared/models/FormInterfaces/other-details.ts
export interface OtherDetails {
  criminalProceedings: boolean;
  criminalConvictions: boolean;
  refusedPassport: boolean;
  impoundedPassport: boolean;
  revokedPassport: boolean;
  grantedCitizenship: boolean;
  heldForeignPassport: boolean;
  surrenderedIndianPassport: boolean;
  appliedRenunciation: boolean;
  returnedOnEc: boolean;
  deportedFromCountry: boolean;
  repatriatedToIndia: boolean;
  registeredMissionName: string;
  // isOtherDetailsValid: boolean;
}